Fixing linting issues
added disable for interface solidity pragma as causing semver issues. Removed variable assigned to delegateCall return for deBridgeCrossChainSend fn.

+++ src/diamond/interfaces/IWethGate.sol
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: BUSL-1.1
+// solhint-disable-next-line
+
 pragma solidity 0.8.7;
 
 interface IWethGate {
     /// @dev Transfer assets to a receiver.
     /// @param receiver This address will receive a transfer.
     /// @param wad Amount in wei
-    function withdraw(address receiver, uint wad) external;
+    function withdraw(address receiver, uint256 wad) external;
 }
